backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Data validation ────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0

# ── Database ───────────────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0
greenlet==3.0.3

# ── Auth ───────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0

# ── AI / LangChain / LangGraph ─────────────────────────────────────
langchain>=0.2.5
langchain-core>=0.2.22
langchain-groq>=0.1.6
langchain-community>=0.2.5
langgraph>=0.1.13

# ── OCR (lightweight, CPU-only) ────────────────────────────────────
pytesseract==0.3.10
Pillow==10.3.0

# ── Utilities ──────────────────────────────────────────────────────
python-dotenv==1.0.1
aiofiles==23.2.1
python-dateutil==2.9.0
pytz==2024.1
numpy==1.26.4
